pyproject.toml
[tool.poetry]
name = "un_yaml"
version = "0.2.0"
description = "Universal YAML: replace repetitive code with your own Domain-Specific un-Language (DSuL)"
authors = ["Ernest Prabhakar <ernest@quiltdata.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "^3.7.0"
trio = "^0.22.0"
asyncclick = "^8.1.3.4"
typing-extensions = "^4.5.0"
pyyaml = "^6.0"
types-pyyaml = "^6.0.12.10"

[tool.poetry.group.dev.dependencies]
pytest-coverage = "^0.0"
pytest-watcher = "^0.2.6"
pytest-asyncio = "^0.21.0"
pytest-trio = "^0.8.0"
pytest-codeblocks = "^0.16.1"
ipykernel = "^6.23.1"
mypy = "^1.3.0"
types-pyyaml = "^6.0.12.10"
types-tzlocal = "^5.0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
